Go home

+++ Java Web/Intellij IDEA/WhatProject/service/src/main/java/com/coppco/service/Impl/DeptServiceImpl.java
@@ -7,6 +7,7 @@ import com.coppco.utils.Page;
 import com.coppco.utils.UtilFuns;
 import org.springframework.context.annotation.Scope;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 
 import javax.annotation.Resource;
@@ -16,6 +17,7 @@ import java.util.List;
 
 @Service(value = "deptService")
 @Scope(value = "prototype")
+@Transactional//事务
 public class DeptServiceImpl implements DeptService {
 
 	@Resource(name = "baseDao")
@@ -35,9 +37,17 @@ public class DeptServiceImpl implements DeptService {
 	}
 
 	public void saveOrUpdate(Dept entity) {
+		//1.调用业务方法，实现保存
+		if (entity.getParent().getId() == null || entity.getParent().getId().length() == 0) {
+			entity.setParent(null);
+		}
 		if(UtilFuns.isEmpty(entity.getId())){
 			//新增
 			entity.setState(1);//1启用  0停用  默认为启用
+		} else  {
+			if (entity.getId().equals(entity.getParent().getId())) {
+				entity.setParent(null);
+			}
 		}
 		baseDao.saveOrUpdate(entity);
 	}
@@ -47,11 +57,28 @@ public class DeptServiceImpl implements DeptService {
 	}
 
 	public void deleteById(Class<Dept> entityClass, Serializable id) {
-		baseDao.deleteById(entityClass, id);
+
+		//删除父部门下面的子部门
+		String sql = "from Dept where parent.id = ?";
+		List<Dept> deptList = baseDao.find(sql, entityClass, new Object[]{id});
+
+		if (deptList != null && deptList.size() > 0) {
+			for (Dept dept : deptList) {
+				deleteById(entityClass, dept.getId());//递归调用
+			}
+		}
+		//删除父部门
+//		baseDao.deleteById(entityClass, id);
+		Dept dept = baseDao.get(entityClass, id);
+		if (dept != null) {
+			baseDao.deleteById(entityClass, id);
+		}
 	}
 
 	public void delete(Class<Dept> entityClass, Serializable[] ids) {
-		baseDao.delete(entityClass, ids);
+		for (Serializable id : ids) {
+			this.deleteById(entityClass, id);
+		}
 	}
 
 }
